Factor out ExecutablePlugin.
